refactor out leaderboard categories
